requirements.txt
Flask==2.1.2
azure-storage-blob==12.10.0
python-dotenv==0.19.2
werkzeug==2.1.2
